+++ Candidates_Project/Candidates.Services/CandidateSkillService.cs
@@ -22,43 +22,67 @@ namespace Candidates.Services
         }
         public async void CreateAsync(CandidateSkillDTO candidateSkillDTO)
         {
-            await _context.Database.EnsureCreatedAsync();
-            var candidateskill = Mapper.Map<CandidateSkillDTO, CandidateSkill>(candidateSkillDTO);
-            await _context.CandidateSkills.AddAsync(candidateskill);
-            await _context.SaveChangesAsync();
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                await _context.Database.EnsureCreatedAsync();
+                var candidateskill = Mapper.Map<CandidateSkillDTO, CandidateSkill>(candidateSkillDTO);
+                await _context.CandidateSkills.AddAsync(candidateskill);
+                await _context.SaveChangesAsync();
+            }
         }
         public async void UpdateAsync(CandidateSkillDTO candidateSkillDTO)
         {
-            var candidateskill = Mapper.Map<CandidateSkillDTO, CandidateSkill>(candidateSkillDTO);
-            _context.CandidateSkills.Update(candidateskill);
-            await _context.SaveChangesAsync();
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var candidateskill = Mapper.Map<CandidateSkillDTO, CandidateSkill>(candidateSkillDTO);
+                _context.CandidateSkills.Update(candidateskill);
+                await _context.SaveChangesAsync();
+            }
         }
         public async void RemoveAsync(int skillID, int candidateID)
         {
-            var candidateSkill = await _context.CandidateSkills.FindAsync(skillID, candidateID);
-            if (candidateSkill != null)
+            if (await _context.Database.EnsureCreatedAsync())
             {
-                _context.CandidateSkills.Remove(candidateSkill);
-                await _context.SaveChangesAsync();
+                var candidateSkill = await _context.CandidateSkills.FindAsync(skillID, candidateID);
+                if (candidateSkill != null)
+                {
+                    _context.CandidateSkills.Remove(candidateSkill);
+                    await _context.SaveChangesAsync();
+                }
             }
         }
         public async Task<CandidateSkillDTO> GetAsync( int skillID, int candidateID)
         {
-            var candidateSkill = await _context.CandidateSkills.FindAsync(skillID, candidateID);
-            var candidateSkillDTO = Mapper.Map<CandidateSkill, CandidateSkillDTO>(candidateSkill);
-            return candidateSkillDTO;
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var candidateSkill = await _context.CandidateSkills.FindAsync(skillID, candidateID);
+                var candidateSkillDTO = Mapper.Map<CandidateSkill, CandidateSkillDTO>(candidateSkill);
+                return candidateSkillDTO;
+            }
+            else
+            {
+                return null;
+            }
         }
         public PageResponse<CandidateSkillDTO> Get(QuerySettings settings)
         {
             var response = new PageResponse<CandidateSkillDTO>();
-            if ((settings.Page - 1) * settings.PageSize + settings.PageSize <= _context.CandidateSkills.Count())
+            if (_context.Database.EnsureCreated())
             {
-                IEnumerable<CandidateSkill> candidatesSkillsPage = _context.CandidateSkills.Skip((settings.Page - 1) * settings.PageSize).Take(settings.PageSize);
-                var candidatesSkillsPageDTO = Mapper.Map<IEnumerable<CandidateSkill>, IEnumerable<CandidateSkillDTO>>(candidatesSkillsPage);
-                response.List = candidatesSkillsPageDTO;
-                response.PageCount = _context.CandidateSkills.Count() / settings.PageSize;
-                response.ItemCount = _context.CandidateSkills.Count();
-               // response.Message = "Ok";
+                if ((settings.Page - 1) * settings.PageSize + settings.PageSize <= _context.CandidateSkills.Count())
+                {
+                    IEnumerable<CandidateSkill> candidatesSkillsPage = _context.CandidateSkills.Skip((settings.Page - 1) * settings.PageSize).Take(settings.PageSize);
+                    var candidatesSkillsPageDTO = Mapper.Map<IEnumerable<CandidateSkill>, IEnumerable<CandidateSkillDTO>>(candidatesSkillsPage);
+                    response.List = candidatesSkillsPageDTO;
+                    response.PageCount = _context.CandidateSkills.Count() / settings.PageSize;
+                    response.ItemCount = _context.CandidateSkills.Count();
+                    // response.Message = "Ok";
+                }
+                else
+                {
+                    //response.Message = $" Incorrect page or item count, max item count: { _context.CandidateSkills.Count() }";
+                    response.ItemCount = _context.CandidateSkills.Count();
+                }
             }
             else
             {
