+++ Candidates_Project/Candidates.Services/HighSchoolService.cs
@@ -22,47 +22,68 @@ namespace Candidates.Services
         }
         public async void CreateAsync(HighSchoolShortDTO highSchoolDTO)
         {
-            _context.Database.EnsureCreated();
-            var highSchool = Mapper.Map<HighSchoolShortDTO, HighSchool>(highSchoolDTO);
-            await _context.HighSchools.AddAsync(highSchool);
-            await _context.SaveChangesAsync();
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var highSchool = Mapper.Map<HighSchoolShortDTO, HighSchool>(highSchoolDTO);
+                await _context.HighSchools.AddAsync(highSchool);
+                await _context.SaveChangesAsync();
+            }
         }
         public async void UpdateAsync( HighSchoolDTO highSchoolDTO)
         {
-            var highSchool = Mapper.Map<HighSchoolDTO, HighSchool>(highSchoolDTO);
-            _context.HighSchools.Update(highSchool);
-            await _context.SaveChangesAsync();
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var highSchool = Mapper.Map<HighSchoolDTO, HighSchool>(highSchoolDTO);
+                _context.HighSchools.Update(highSchool);
+                await _context.SaveChangesAsync();
+            }
         }
         public async void RemoveAsync(int id)
         {
-            var highschool = await _context.HighSchools.FindAsync(id);
-            if (highschool != null)
-            { 
-                _context.HighSchools.Remove(highschool);
-                await _context.SaveChangesAsync();
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var highschool = await _context.HighSchools.FindAsync(id);
+                if (highschool != null)
+                {
+                    _context.HighSchools.Remove(highschool);
+                    await _context.SaveChangesAsync();
+                }
             }
         }
         public async Task<HighSchoolDTO> GetAsync(int id)
         {
-            var highschool = await _context.HighSchools.FindAsync(id);
-            var highschoolDTO = Mapper.Map<HighSchool,HighSchoolDTO>(highschool);
-            return highschoolDTO;
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var highschool = await _context.HighSchools.FindAsync(id);
+                var highschoolDTO = Mapper.Map<HighSchool, HighSchoolDTO>(highschool);
+                return highschoolDTO;
+            }
+            else
+                return null;
         }
         public PageResponse<HighSchoolDTO> Get(QuerySettings settings)
         {
             var response = new PageResponse<HighSchoolDTO>();
-            if ((settings.Page - 1) * settings.PageSize + settings.PageSize <= _context.HighSchools.Count())
+            if (_context.Database.EnsureCreated())
             {
-                IEnumerable<HighSchool> highSchoolsPage = _context.HighSchools.Skip((settings.Page - 1) * settings.PageSize).Take(settings.PageSize);
-                var highSchoolsPageDTO = Mapper.Map<IEnumerable<HighSchool>, IEnumerable<HighSchoolDTO>>(highSchoolsPage);
-                response.List = highSchoolsPageDTO;
-                response.PageCount = _context.HighSchools.Count() / settings.PageSize;
-                response.ItemCount = _context.HighSchools.Count();
-                //response.Message = "Ok";
+                
+                if ((settings.Page - 1) * settings.PageSize + settings.PageSize <= _context.HighSchools.Count())
+                {
+                    IEnumerable<HighSchool> highSchoolsPage = _context.HighSchools.Skip((settings.Page - 1) * settings.PageSize).Take(settings.PageSize);
+                    var highSchoolsPageDTO = Mapper.Map<IEnumerable<HighSchool>, IEnumerable<HighSchoolDTO>>(highSchoolsPage);
+                    response.List = highSchoolsPageDTO;
+                    response.PageCount = _context.HighSchools.Count() / settings.PageSize;
+                    response.ItemCount = _context.HighSchools.Count();
+                    //response.Message = "Ok";
+                }
+                else
+                {
+                    // response.Message = $" Incorrect page or item count, max item count: { _context.HighSchools.Count() }";
+                    response.ItemCount = _context.HighSchools.Count();
+                }
             }
             else
             {
-               // response.Message = $" Incorrect page or item count, max item count: { _context.HighSchools.Count() }";
                 response.ItemCount = _context.HighSchools.Count();
             }
             return response;
