+++ Candidates_Project/Candidates.Services/LanguageService.cs
@@ -22,47 +22,70 @@ namespace Candidates.Services
         }
         public async void CreateAsync(LanguageShortDTO languageDTO)
         {
-            await _context.Database.EnsureCreatedAsync();
-            var language = Mapper.Map<LanguageShortDTO, Language>(languageDTO);
-            await _context.Languages.AddAsync(language);
-            _context.SaveChanges();
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var language = Mapper.Map<LanguageShortDTO, Language>(languageDTO);
+                await _context.Languages.AddAsync(language);
+                _context.SaveChanges();
+            }
         }
         public async void UpdateAsync(LanguageDTO languageDTO)
         {
-            var language = Mapper.Map<LanguageDTO, Language>(languageDTO);
-            _context.Languages.Update(language);
-            await _context.SaveChangesAsync();
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var language = Mapper.Map<LanguageDTO, Language>(languageDTO);
+                _context.Languages.Update(language);
+                await _context.SaveChangesAsync();
+            }
         }
         public async void RemoveAsync(int id)
         {
-            var language = await _context.Languages.FindAsync(id);
-            if (language != null)
+            if (await _context.Database.EnsureCreatedAsync())
             {
-                _context.Languages.Remove(language);
-                await _context.SaveChangesAsync();
+                var language = await _context.Languages.FindAsync(id);
+                if (language != null)
+                {
+                    _context.Languages.Remove(language);
+                    await _context.SaveChangesAsync();
+                }
             }
         }
         public async Task<LanguageDTO> GetAsync( int id)
         {
-            var language = await _context.Languages.FindAsync(id);
-            var languageDTO = Mapper.Map<Language, LanguageDTO>(language);
-            return languageDTO;
+            if (await _context.Database.EnsureCreatedAsync())
+            {
+                var language = await _context.Languages.FindAsync(id);
+                var languageDTO = Mapper.Map<Language, LanguageDTO>(language);
+                return languageDTO;
+            }
+            else
+            {
+                return null;
+            }
         }
         public PageResponse<LanguageDTO> Get(QuerySettings settings)
         {
             var response = new PageResponse<LanguageDTO>();
-            if ((settings.Page - 1) * settings.PageSize + settings.PageSize <= _context.Languages.Count())
+            if ( _context.Database.EnsureCreated())
             {
-                IEnumerable<Language> languagesPage = _context.Languages.Skip((settings.Page - 1) * settings.PageSize).Take(settings.PageSize);
-                var languagesPageDTO = Mapper.Map<IEnumerable<Language>, IEnumerable<LanguageDTO>>(languagesPage);
-                response.List = languagesPageDTO;
-                response.PageCount = _context.Languages.Count() / settings.PageSize;
-                response.ItemCount = _context.Languages.Count();
-               // response.Message = "Ok";
+                if ((settings.Page - 1) * settings.PageSize + settings.PageSize <= _context.Languages.Count())
+                {
+                    IEnumerable<Language> languagesPage = _context.Languages.Skip((settings.Page - 1) * settings.PageSize).Take(settings.PageSize);
+                    var languagesPageDTO = Mapper.Map<IEnumerable<Language>, IEnumerable<LanguageDTO>>(languagesPage);
+                    response.List = languagesPageDTO;
+                    response.PageCount = _context.Languages.Count() / settings.PageSize;
+                    response.ItemCount = _context.Languages.Count();
+                    // response.Message = "Ok";
+                }
+                else
+                {
+                    // response.Message = $" Incorrect page or item count, max item count: { _context.Languages.Count() }";
+                    response.ItemCount = _context.Languages.Count();
+                }
             }
             else
             {
-               // response.Message = $" Incorrect page or item count, max item count: { _context.Languages.Count() }";
+                // response.Message = $" Incorrect page or item count, max item count: { _context.Languages.Count() }";
                 response.ItemCount = _context.Languages.Count();
             }
             return response;
